replace uuid implementation
